# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookshelf API",
        "description": "API documentation for Bookshelf application",
        "contact": {
            "email": "support@test.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bookshelves/{bookshelfId}/books": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get all books in a bookshelf",
                "operationId": "456ebb28e3bc14fb681fa9d06e41d374",
                "parameters": [
                    {
                        "name": "bookshelfId",
                        "in": "path",
                        "description": "ID of the bookshelf",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Bookshelf not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Create a new book in a bookshelf",
                "operationId": "8faffa70ccfdca53b9f35bd5d99b170e",
                "parameters": [
                    {
                        "name": "bookshelfId",
                        "in": "path",
                        "description": "ID of the bookshelf",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "published_year"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "published_year": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully"
                    },
                    "404": {
                        "description": "Bookshelf not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/bookshelves/{bookshelfId}/books/{bookId}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get a specific book in a bookshelf",
                "operationId": "7ef75f011f0c6d50ad69ff6f8d42c2e6",
                "parameters": [
                    {
                        "name": "bookshelfId",
                        "in": "path",
                        "description": "ID of the bookshelf",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Bookshelf or Book not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Book"
                ],
                "summary": "Update a specific book in a bookshelf",
                "operationId": "ad675061935b0e3bca61c2ef9ac0927a",
                "parameters": [
                    {
                        "name": "bookshelfId",
                        "in": "path",
                        "description": "ID of the bookshelf",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "published_year"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "published_year": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated successfully"
                    },
                    "404": {
                        "description": "Bookshelf or Book not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Book"
                ],
                "summary": "Delete a specific book in a bookshelf",
                "operationId": "2a701c014e25cfa4ad387eedd016fbb2",
                "parameters": [
                    {
                        "name": "bookshelfId",
                        "in": "path",
                        "description": "ID of the bookshelf",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book deleted successfully"
                    },
                    "404": {
                        "description": "Bookshelf or Book not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/books/search": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Search for books by title or author",
                "operationId": "a50e12f6ebf00d1971a30d2fe743509c",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title of the book",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "Author of the book",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearerAuth",
                "bearerFormat": "JWT"
            }
        }
    }
}